requirements.txt
FastAPI
langchain
neo4j
redis
streamlit
openai
uvicorn
PyPDF2
python-dotenv
python-multipart
langchain-core
langchain-openai
langchain-neo4j>=0.1.0
tqdm
langchain-community
numpy
langchain-huggingface
langchain-experimental
chardet
sentence-transformers
torch
pypdf
langchain-text-splitters
spacy>=3.0.0,<4.0.0
scispacy==0.5.3
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz